refactor: Remove functional from the new class titles and rename the old physics modules to legacy
