+++ src/serac/infrastructure/tests/serac_error_handling.cpp
@@ -12,7 +12,7 @@
 #include "serac/infrastructure/cli.hpp"
 #include "serac/infrastructure/initialize.hpp"
 #include "serac/mesh/mesh_utils.hpp"
-#include "serac/physics/thermal_conduction.hpp"
+#include "serac/physics/thermal_conduction_legacy.hpp"
 #include "serac/physics/boundary_conditions/boundary_condition.hpp"
 #include "serac/numerics/equation_solver.hpp"
 #include "serac/physics/state/state_manager.hpp"
@@ -47,8 +47,8 @@ TEST(SeracErrorHandling, EquationSolverAmgxNotAvailable)
 #ifndef MFEM_USE_SUNDIALS
 TEST(SeracErrorHandling, EquationSolverKinsolNotAvailable)
 {
-  auto lin_options             = ThermalConduction::defaultLinearOptions();
-  auto nonlin_options          = ThermalConduction::defaultNonlinearOptions();
+  auto lin_options             = ThermalConductionLegacy::defaultLinearOptions();
+  auto nonlin_options          = ThermalConductionLegacy::defaultNonlinearOptions();
   nonlin_options.nonlin_solver = NonlinearSolver::KINFullStep;
   EXPECT_THROW(EquationSolver(MPI_COMM_WORLD, lin_options, nonlin_options), SlicErrorException);
 }
