+++ src/serac/physics/materials/parameterized_solid_functional_material.hpp
@@ -15,7 +15,7 @@
 #include "serac/numerics/functional/functional.hpp"
 #include "serac/physics/materials/solid_functional_material.hpp"
 
-/// SolidFunctional helper data types
+/// Solid helper data types
 namespace serac::solid_mechanics {
 
 /**
