+++ src/serac/physics/tests/serac_solid_functional_shape.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/physics/solid_functional.hpp"
+#include "serac/physics/solid.hpp"
 #include "serac/physics/materials/solid_functional_material.hpp"
 #include "serac/physics/materials/parameterized_solid_functional_material.hpp"
 
@@ -97,7 +97,7 @@ void shape_test(GeometricNonlinearities geo_nonlin)
     user_defined_shape_displacement.project(shape_coef);
 
     // Construct a functional-based solid mechanics solver including references to the shape velocity field.
-    SolidFunctional<p, dim> solid_solver(options, geo_nonlin, "solid_functional", ShapeDisplacement::On);
+    Solid<p, dim> solid_solver(options, geo_nonlin, "solid_functional", ShapeDisplacement::On);
 
     // Set the initial displacement and boundary condition
     solid_solver.setDisplacementBCs(ess_bdr, bc);
@@ -139,7 +139,7 @@ void shape_test(GeometricNonlinearities geo_nonlin)
     *mesh_nodes += user_defined_shape_displacement.gridFunction();
 
     // Construct a functional-based solid mechanics solver including references to the shape velocity field.
-    SolidFunctional<p, dim> solid_solver_no_shape(options, geo_nonlin, "solid_functional");
+    Solid<p, dim> solid_solver_no_shape(options, geo_nonlin, "solid_functional");
 
     mfem::VisItDataCollection visit_dc("pure_version", const_cast<mfem::ParMesh*>(&solid_solver_no_shape.mesh()));
     visit_dc.RegisterField("displacement", &solid_solver_no_shape.displacement().gridFunction());
