+++ src/serac/physics/tests/serac_thermal_functional.cpp
@@ -12,7 +12,7 @@
 
 #include "serac/serac_config.hpp"
 #include "serac/mesh/mesh_utils.hpp"
-#include "serac/physics/thermal_conduction_functional.hpp"
+#include "serac/physics/thermal_conduction.hpp"
 #include "serac/physics/materials/thermal_functional_material.hpp"
 #include "serac/physics/materials/parameterized_thermal_functional_material.hpp"
 #include "serac/physics/state/state_manager.hpp"
@@ -44,7 +44,7 @@ void functional_test_static(double expected_temp_norm)
   std::set<int> ess_bdr = {1};
 
   // Construct a functional-based thermal conduction solver
-  ThermalConductionFunctional<p, dim> thermal_solver(Thermal::defaultQuasistaticOptions(), "thermal_functional");
+  ThermalConduction<p, dim> thermal_solver(Thermal::defaultQuasistaticOptions(), "thermal_functional");
 
   tensor<double, dim, dim> cond;
 
@@ -114,7 +114,7 @@ void functional_test_dynamic(double expected_temp_norm)
   std::set<int> ess_bdr = {1};
 
   // Construct a functional-based thermal conduction solver
-  ThermalConductionFunctional<p, dim> thermal_solver(Thermal::defaultDynamicOptions(), "thermal_functional");
+  ThermalConduction<p, dim> thermal_solver(Thermal::defaultDynamicOptions(), "thermal_functional");
 
   // Define an isotropic conductor material model
   Thermal::LinearIsotropicConductor mat(0.5, 0.5, 0.5);
@@ -204,7 +204,7 @@ TEST(ThermalFunctional, ParameterizedMaterial)
   // Note that we now include an extra template parameter indicating the finite element space for the parameterized
   // field, in this case the thermal conductivity. We also pass an array of finite element states for each of the
   // requested parameterized fields.
-  ThermalConductionFunctional<p, dim, Parameters<H1<1> > > thermal_solver(Thermal::defaultQuasistaticOptions(),
+  ThermalConduction<p, dim, Parameters<H1<1> > > thermal_solver(Thermal::defaultQuasistaticOptions(),
                                                                           "thermal_functional");
   thermal_solver.setParameter(user_defined_conductivity, 0);
